pyproject.toml
[project]
name = "talk-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.1",
    "fastrtc>=0.0.33",
    "google>=3.0.0",
    "google-api-python-client>=2.187.0",
    "google-cloud>=0.34.0",
    "google-genai>=1.50.0",
    "googleapis-common-protos>=1.72.0",
    "gradio>=5.49.1",
    "gtts>=2.5.4",
    "numpy>=2.3.4",
    "python-socketio>=5.14.3",
    "sounddevice>=0.5.3",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]
